+++ modules/exams/service/exam.service.js
@@ -23,17 +23,35 @@ const getExam=async (filters={})=>{
     return await query.select('*');
 }
 
-const deleteExam=async(exam_id)=>{
-    const exam=await db('exams').where('exam_id',exam_id).first();
-    if(!exam)return "Exam not found";
-    const filePath = path.join(__dirname, '..', '..', '..', 'public', 'uploads', 'exam', path.basename(exam.exam_path));
-    if (fs.existsSync(filePath)) {
-        fs.unlinkSync(filePath);
+const deleteExam = async (exam_id, admin_id) => {
+    const exam = await db('exams').where('exam_id', exam_id).first();
+    if (!exam) return "Exam not found";
+    if (!exam.exam_path || typeof exam.exam_path !== 'string') {
+        throw new Error("exam_path is null or invalid.");
     }
-    await db('favorites').where({content_type:'exam',content_id:exam_id}).delete();
-    await db('exams').where('exam_id',exam_id).delete();
+    const fileName = path.basename(exam.exam_path);
+    const oldPath = path.join(__dirname, '..', '..', '..', 'public', exam.exam_path);
+    const archivePath = path.join(__dirname, '..', '..', '..', 'public', 'uploads', 'archive', 'exam', fileName);
+    await db('favorites').where({ content_type: 'exam', content_id: exam_id }).delete();
+    if (fs.existsSync(oldPath)) {
+        fs.mkdirSync(path.dirname(archivePath), { recursive: true });
+        fs.renameSync(oldPath, archivePath);
+    } else {
+        console.warn("⚠️ الملف غير موجود في المسار:", oldPath);
+    }
+    await db('archives').insert({
+        content_id: exam_id,
+        content_type: 'exam',
+        file_path: `archive/exam/${fileName}`,
+        deleted_by: admin_id,
+        original_data: JSON.stringify(exam)
+    });
+
+    await db('exams').where('exam_id', exam_id).delete();
+
     return true;
-}
+};
+
 
 const updateExam=async(exam_id,data)=>{
     return await db('exams').where('exam_id',exam_id).update(data);
